+++ xmpp-chat-service/src/main/java/com/algomeet/xmpp/chatservice/routing/chat/OfflineMessageHandler.java
@@ -82,7 +82,7 @@ public class OfflineMessageHandler {
                     return localStanzaDispatcher.dispatchLocally(userKey, msg.getFrom(), xmlWithDelay)
                             // If dispatchLocally returns a Mono<Boolean>, intercept the result here
                             .doOnNext(isSuccess -> {
-                                if (Boolean.TRUE.equals(isSuccess)) {
+                                if (Boolean.TRUE.equals(isSuccess) && msg.getIsAckStanza()) {
                                 	// Delete if record is ACK stanza
                                 	offlineMessageRepository.deleteByIdAndIsAckStanzaTrue(msg.getId()).subscribe();
                                  }
